Track only static data in scheduling PostgresResultsCell
A RequestRecord provides the state of a scheduling_request row at the
time it is requested. A PostgresResultCell's data is likely to persist
long after its creation, and thus the data in the RequestRecord it was
created from may very well be updated before it is used again. To
avoid keeping around stale data, we only keep references to the data
that identifies the request and its associated analysis, which should
never change.